+++ apps/common/hooks/useMultichainBalances.ts
@@ -19,7 +19,7 @@ import type {ContractFunctionConfig} from 'viem';
 import type {Connector} from 'wagmi';
 import type {TAddress, TDict, TNDict} from '@yearn-finance/web-lib/types';
 import type {TDefaultStatus} from '@yearn-finance/web-lib/types/hooks';
-import type {TYDaemonPrices} from '@yearn-finance/web-lib/utils/schemas/yDaemonPricesSchema';
+import type {TYDaemonPricesChain} from '@common/schemas/yDaemonPricesSchema';
 import type {TChainTokens, TToken} from '@common/types/types';
 
 /* 🔵 - Yearn Finance **********************************************************
@@ -36,7 +36,7 @@ export type TUseBalancesTokens = {
 export type TUseBalancesReq = {
 	key?: string | number;
 	tokens: TUseBalancesTokens[];
-	prices?: TYDaemonPrices;
+	prices?: TYDaemonPricesChain;
 	effectDependencies?: DependencyList;
 	provider?: Connector;
 };
@@ -71,7 +71,7 @@ async function performCall(
 	chainID: number,
 	calls: ContractFunctionConfig[],
 	tokens: TUseBalancesTokens[],
-	prices?: TYDaemonPrices
+	prices?: TYDaemonPricesChain
 ): Promise<[TDict<TToken>, Error | undefined]> {
 	const _data: TDict<TToken> = {};
 	const results = await multicall({
@@ -84,7 +84,7 @@ async function performCall(
 		const {address, decimals: injectedDecimals, name: injectedName, symbol: injectedSymbol} = element;
 		const balanceOf = decodeAsBigInt(results[rIndex++]);
 		const decimals = decodeAsNumber(results[rIndex++]) || injectedDecimals || 18;
-		const rawPrice = toBigInt(prices?.[address]);
+		const rawPrice = toBigInt(prices?.[chainID]?.[address]);
 		let symbol = decodeAsString(results[rIndex++]) || injectedSymbol || '';
 		let name = decodeAsString(results[rIndex++]) || injectedName || '';
 		if (isEth(address)) {
@@ -114,7 +114,7 @@ async function getBalances(
 	chainID: number,
 	address: TAddress,
 	tokens: TUseBalancesTokens[],
-	prices?: TYDaemonPrices
+	prices?: TYDaemonPricesChain
 ): Promise<[TDict<TToken>, Error | undefined]> {
 	let result: TDict<TToken> = {};
 	const calls: ContractFunctionConfig[] = [];
@@ -370,7 +370,7 @@ export function useBalances(props?: TUseBalancesReq): TUseBalancesRes {
 				const chainID = Number(chainIDStr);
 				for (const address of Object.keys(newData[chainID])) {
 					const tokenAddress = toAddress(address);
-					const rawPrice = toBigInt(props?.prices?.[tokenAddress]);
+					const rawPrice = toBigInt(props?.prices?.[chainID]?.[tokenAddress]);
 					if (!newData[chainID]) {
 						newData[chainID] = {};
 					}